+++ packages/opentelemetry-core/test/common/attributes.test.ts
@@ -14,13 +14,28 @@
  * limitations under the License.
  */
 
+import { diag } from '@opentelemetry/api';
+import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
 import * as assert from 'assert';
+import * as sinon from 'sinon';
 import {
   isAttributeValue,
+  sanitizeAttribute,
   sanitizeAttributes,
 } from '../../src/common/attributes';
 
 describe('attributes', () => {
+  const warnStub = sinon.fake();
+
+  beforeEach(() => {
+    diag.setLogger({
+      warn: warnStub,
+    } as any);
+
+    // diag.warn is used when the logger is set
+    warnStub.resetHistory();
+  });
+
   describe('#isAttributeValue', () => {
     it('should allow primitive values', () => {
       assert.ok(isAttributeValue(0));
@@ -101,6 +116,79 @@ describe('attributes', () => {
       assert.strictEqual(attributes.str, 'unmodified');
       assert.ok(Array.isArray(attributes.arr));
       assert.strictEqual(attributes.arr[0], 'unmodified');
+    });
+
+    describe('value sanitizers', () => {
+      describe('http.url', () => {
+        it('should remove username and password from the url', () => {
+          const out = sanitizeAttribute(
+            SemanticAttributes.HTTP_URL,
+            'http://user:pass@host:9000/path?query#fragment'
+          );
+
+          assert.strictEqual(out, 'http://host:9000/path?query#fragment');
+          assert.ok(warnStub.notCalled, 'should not log warning');
+        });
+
+        it('should return the same url', () => {
+          const out = sanitizeAttribute(
+            SemanticAttributes.HTTP_URL,
+            'http://host:9000/path?query#fragment'
+          );
+
+          assert.strictEqual(out, 'http://host:9000/path?query#fragment');
+          assert.ok(warnStub.notCalled, 'should not log warning');
+        });
+
+        it('should return the input string when the value is not a valid url', () => {
+          const out = sanitizeAttribute(
+            SemanticAttributes.HTTP_URL,
+            'invalid url'
+          );
+
+          assert.strictEqual(out, 'invalid url');
+          assert.ok(
+            warnStub.calledWithExactly(
+              `Invalid attribute value set for key: ${SemanticAttributes.HTTP_URL}. Unable to sanitize invalid URL.`
+            )
+          );
+        });
+
+        it('should return the input when the value is a number', () => {
+          const out = sanitizeAttribute(SemanticAttributes.HTTP_URL, 27);
+
+          assert.strictEqual(out, 27);
+          assert.ok(
+            warnStub.calledWithExactly(
+              `Invalid attribute value set for key: ${SemanticAttributes.HTTP_URL}. Unable to sanitize number value.`
+            )
+          );
+        });
+
+        it('should return the input when the value is boolean', () => {
+          const out = sanitizeAttribute(SemanticAttributes.HTTP_URL, false);
+
+          assert.strictEqual(out, false);
+          assert.ok(
+            warnStub.calledWithExactly(
+              `Invalid attribute value set for key: ${SemanticAttributes.HTTP_URL}. Unable to sanitize boolean value.`
+            )
+          );
+        });
+
+        it('should return the input when an array is supplied', () => {
+          const out = sanitizeAttribute(SemanticAttributes.HTTP_URL, [
+            'http://host/path?query#fragment',
+          ]);
+
+          assert.deepStrictEqual(out, ['http://host/path?query#fragment']);
+          assert.ok(
+            warnStub.calledWithExactly(
+              `Invalid attribute value set for key: ${SemanticAttributes.HTTP_URL}. Unable to sanitize array value.`
+            )
+          );
+        });
+      });
     });
   });
 });